Backend/requirements.txt
asgiref==3.7.2
Django==5.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
psycopg==3.1.16
psycopg-binary==3.1.16
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
